MAINT: Move testing tools from nose_tools to testing directory.
These tools do not introduce a nose dependency unless the NoseTester is
instanciated, the nose imports are lazy. Consequently we do not need to
keep them separated out.
