+++ numpy/testing/pytestdecorators.py
