+++ numpy/testing/setup.py
@@ -4,10 +4,8 @@ from __future__ import division, print_function
 
 def configuration(parent_package='',top_path=None):
     from numpy.distutils.misc_util import Configuration
-    config = Configuration('testing', parent_package, top_path)
 
-    config.add_subpackage('nose_tools')
-    config.add_subpackage('pytest_tools')
+    config = Configuration('testing', parent_package, top_path)
     config.add_data_dir('tests')
     return config
